requirements.txt
streamlit>=1.20
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
openai>=0.27.0
transformers>=4.30.0
PyPDF2>=3.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
numpy>=1.24
pandas>=2.0
